Use the new retryable API gear in artifact uploads.
