+++ buildkite/artifact.go
@@ -1,21 +1,19 @@
 package buildkite
 
 import (
-	"crypto/sha1"
-	"errors"
 	"fmt"
 	"github.com/buildkite/agent/buildkite/logger"
-	"io"
 	"mime"
 	"net/url"
-	"os"
 	"path/filepath"
-	"strings"
 )
 
 type Artifact struct {
 	// The ID of the artifact
 	ID string `json:"id,omitempty"`
+
+	// The ID of the job
+	JobID string
 
 	// The current state of the artifact. Default is "new"
 	State string `json:"state,omitempty"`
@@ -60,10 +58,17 @@ type Artifact struct {
 		// Data that should be sent along with the upload
 		Data map[string]string
 	}
+
+	// The API used for communication
+	API API
 }
 
 func (a Artifact) String() string {
 	return fmt.Sprintf("Artifact{ID: %s, Path: %s, URL: %s, AbsolutePath: %s, GlobPath: %s, FileSize: %d, Sha1Sum: %s}", a.ID, a.Path, a.URL, a.AbsolutePath, a.GlobPath, a.FileSize, a.Sha1Sum)
+}
+
+func (a *Artifact) Update() error {
+	return a.API.Put("/jobs/"+a.JobID+"/artifacts/"+a.ID, &a, a)
 }
 
 func (a Artifact) MimeType() string {
@@ -77,210 +82,11 @@ func (a Artifact) MimeType() string {
 	}
 }
 
-func (c *Client) ArtifactUpdate(jobID string, artifact Artifact) (*Artifact, error) {
-	// Create a new instance of a artifact that will be populated
-	// with the updated data by the client
-	var updatedArtifact Artifact
-
-	// Return the job.
-	return &updatedArtifact, c.Put(&updatedArtifact, "jobs/"+jobID+"/artifacts/"+artifact.ID, artifact)
-}
-
-// Sends all the artifacts at once to the Buildkite Agent API. This will allow
-// the UI to show what artifacts will be uploaded. Their state starts out as
-// "new"
-func (c *Client) CreateArtifacts(jobID string, artifacts []*Artifact) ([]Artifact, error) {
-	var createdArtifacts []Artifact
-
-	return createdArtifacts, c.Post(&createdArtifacts, "jobs/"+jobID+"/artifacts", artifacts)
-}
-
 // Searches for artifacts on the build
 func (c *Client) SearchArtifacts(buildId string, searchQuery string, jobQuery string, stateQuery string) ([]Artifact, error) {
 	var foundArtifacts []Artifact
 
 	return foundArtifacts, c.Get(&foundArtifacts, "builds/"+buildId+"/artifacts/search?query="+url.QueryEscape(searchQuery)+"&job="+url.QueryEscape(jobQuery)+"&state="+url.QueryEscape(stateQuery))
-}
-
-func CollectArtifacts(artifactPaths string) (artifacts []*Artifact, err error) {
-	globs := strings.Split(artifactPaths, ";")
-	workingDirectory, _ := os.Getwd()
-
-	for _, glob := range globs {
-		glob = strings.TrimSpace(glob)
-
-		if glob != "" {
-			logger.Debug("Globbing %s for %s", workingDirectory, glob)
-
-			files, err := Glob(workingDirectory, glob)
-			if err != nil {
-				return nil, err
-			}
-
-			for _, file := range files {
-				// Generate an absolute path for the artifact
-				absolutePath, err := filepath.Abs(file)
-				if err != nil {
-					return nil, err
-				}
-
-				fileInfo, err := os.Stat(absolutePath)
-				if fileInfo.IsDir() {
-					logger.Debug("Skipping directory %s", file)
-					continue
-				}
-
-				// Create a relative path (from the workingDirectory) to the artifact, by removing the
-				// first part of the absolutePath that is the workingDirectory.
-				relativePath := strings.Replace(absolutePath, workingDirectory, "", 1)
-
-				// Ensure the relativePath doesn't have a file seperator "/" as the first character
-				relativePath = strings.TrimPrefix(relativePath, string(os.PathSeparator))
-
-				// Build an artifact object using the paths we have.
-				artifact, err := BuildArtifact(relativePath, absolutePath, glob)
-				if err != nil {
-					return nil, err
-				}
-
-				artifacts = append(artifacts, artifact)
-			}
-		}
-	}
-
-	return artifacts, nil
-}
-
-func BuildArtifact(relativePath string, absolutePath string, globPath string) (*Artifact, error) {
-	// Temporarily open the file to get it's size
-	file, err := os.Open(absolutePath)
-	if err != nil {
-		return nil, err
-	}
-	defer file.Close()
-
-	// Grab it's file info (which includes it's file size)
-	fileInfo, err := file.Stat()
-	if err != nil {
-		return nil, err
-	}
-
-	// Generate a sha1 checksum for the file
-	hash := sha1.New()
-	io.Copy(hash, file)
-	checksum := fmt.Sprintf("%x", hash.Sum(nil))
-
-	// Create our new artifact data structure
-	artifact := new(Artifact)
-	artifact.State = "new"
-	artifact.Path = relativePath
-	artifact.AbsolutePath = absolutePath
-	artifact.GlobPath = globPath
-	artifact.FileSize = fileInfo.Size()
-	artifact.Sha1Sum = checksum
-
-	return artifact, nil
-}
-
-func UploadArtifacts(client Client, jobID string, artifacts []*Artifact, destination string) error {
-	var uploader Uploader
-
-	// Determine what uploader to use
-	if destination != "" {
-		if strings.HasPrefix(destination, "s3://") {
-			uploader = new(S3Uploader)
-		} else {
-			return errors.New("Unknown upload destination: " + destination)
-		}
-	} else {
-		uploader = new(FormUploader)
-	}
-
-	// Setup the uploader
-	err := uploader.Setup(destination)
-	if err != nil {
-		return err
-	}
-
-	// Set the URL's of the artifacts based on the uploader
-	for _, artifact := range artifacts {
-		artifact.URL = uploader.URL(artifact)
-	}
-
-	// Create artifacts on buildkite in batches to prevent timeouts with many artifacts
-	var lenArtifacts = len(artifacts)
-	var createdArtifacts = []Artifact{}
-
-	for i := 0; i < lenArtifacts; i += 100 {
-		j := i + 100
-		if lenArtifacts < j {
-			j = lenArtifacts
-		}
-
-		someArtifacts := artifacts[i:j]
-
-		someCreatedArtifacts, err := client.CreateArtifacts(jobID, someArtifacts)
-		if err != nil {
-			return err
-		}
-
-		createdArtifacts = append(createdArtifacts, someCreatedArtifacts...)
-	}
-
-	// Upload the artifacts by spinning up some routines
-	var routines []chan string
-	var concurrency int = 10
-
-	logger.Debug("Spinning up %d concurrent threads for uploads", concurrency)
-
-	count := 0
-	for _, artifact := range createdArtifacts {
-		// Create a channel and append it to the routines array. Once we've hit our
-		// concurrency limit, we'll block until one finishes, then this loop will
-		// startup up again.
-		count++
-		wait := make(chan string)
-		go uploadRoutine(wait, client, jobID, artifact, uploader)
-		routines = append(routines, wait)
-
-		if count >= concurrency {
-			logger.Debug("Maximum concurrent threads running. Waiting.")
-
-			// Wait for all the routines to finish, then reset
-			waitForRoutines(routines)
-			count = 0
-			routines = routines[0:0]
-		}
-	}
-
-	// Wait for any other routines to finish
-	waitForRoutines(routines)
-
-	return nil
-}
-
-func uploadRoutine(quit chan string, client Client, jobID string, artifact Artifact, uploader Uploader) {
-	// Show a nice message that we're starting to upload the file
-	logger.Info("Uploading \"%s\" (%d bytes)", artifact.Path, artifact.FileSize)
-
-	// Upload the artifact and then set the state depending on whether or not
-	// it passed.
-	err := uploader.Upload(&artifact)
-	if err != nil {
-		artifact.State = "error"
-		logger.Error("Error uploading artifact \"%s\": %s", artifact.Path, err)
-	} else {
-		artifact.State = "finished"
-	}
-
-	// Update the state of the artifact on Buildkite
-	_, err = client.ArtifactUpdate(jobID, artifact)
-	if err != nil {
-		logger.Error("Error marking artifact %s as uploaded: %s", artifact.Path, err)
-	}
-
-	// We can notify the channel that this routine has finished now
-	quit <- "finished"
 }
 
 func DownloadArtifacts(artifacts []Artifact, destination string) error {
@@ -305,19 +111,19 @@ func DownloadArtifacts(artifacts []Artifact, destination string) error {
 			logger.Debug("Maximum concurrent threads running. Waiting.")
 
 			// Wait for all the routines to finish, then reset
-			waitForRoutines(routines)
+			waitForDownloadRoutines(routines)
 			count = 0
 			routines = routines[0:0]
 		}
 	}
 
 	// Wait for any other routines to finish
-	waitForRoutines(routines)
+	waitForDownloadRoutines(routines)
 
 	return nil
 }
 
-func waitForRoutines(routines []chan string) {
+func waitForDownloadRoutines(routines []chan string) {
 	for _, r := range routines {
 		<-r
 	}
